Refactor RayTune
