+++ ultralytics/yolo/engine/model.py
@@ -9,7 +9,7 @@ from ultralytics.nn.tasks import (ClassificationModel, DetectionModel, PoseModel
                                   attempt_load_one_weight, guess_model_task, nn, yaml_model_load)
 from ultralytics.yolo.cfg import get_cfg
 from ultralytics.yolo.engine.exporter import Exporter
-from ultralytics.yolo.utils import (DEFAULT_CFG, DEFAULT_CFG_DICT, DEFAULT_CFG_KEYS, LOGGER, NUM_THREADS, RANK, ROOT,
+from ultralytics.yolo.utils import (DEFAULT_CFG, DEFAULT_CFG_DICT, DEFAULT_CFG_KEYS, LOGGER, RANK, ROOT,
                                     callbacks, is_git_dir, yaml_load)
 from ultralytics.yolo.utils.checks import check_file, check_imgsz, check_pip_update_available, check_yaml
 from ultralytics.yolo.utils.downloads import GITHUB_ASSET_STEMS
@@ -387,13 +387,7 @@ class YOLO:
         self._check_is_pytorch_model()
         self.model.to(device)
 
-    def tune(self,
-             data: str,
-             space: dict = None,
-             grace_period: int = 10,
-             gpu_per_trial: int = None,
-             max_samples: int = 10,
-             train_args: dict = None):
+    def tune(self, *args, **kwargs):
         """
         Runs hyperparameter tuning using Ray Tune.
 
@@ -411,66 +405,9 @@ class YOLO:
         Raises:
             ModuleNotFoundError: If Ray Tune is not installed.
         """
-        if train_args is None:
-            train_args = {}
-
-        try:
-            from ultralytics.yolo.utils.tuner import (ASHAScheduler, RunConfig, WandbLoggerCallback, default_space,
-                                                      task_metric_map, tune)
-        except ImportError:
-            raise ModuleNotFoundError("Install Ray Tune: `pip install 'ray[tune]'`")
-
-        try:
-            import wandb
-            from wandb import __version__  # noqa
-        except ImportError:
-            wandb = False
-
-        def _tune(config):
-            """
-            Trains the YOLO model with the specified hyperparameters and additional arguments.
-
-            Args:
-                config (dict): A dictionary of hyperparameters to use for training.
-
-            Returns:
-                None.
-            """
-            self._reset_callbacks()
-            config.update(train_args)
-            self.train(**config)
-
-        if not space:
-            LOGGER.warning('WARNING: search space not provided. Using default search space')
-            space = default_space
-
-        space['data'] = data
-
-        # Define the trainable function with allocated resources
-        trainable_with_resources = tune.with_resources(_tune, {'cpu': NUM_THREADS, 'gpu': gpu_per_trial or 0})
-
-        # Define the ASHA scheduler for hyperparameter search
-        asha_scheduler = ASHAScheduler(time_attr='epoch',
-                                       metric=task_metric_map[self.task],
-                                       mode='max',
-                                       max_t=train_args.get('epochs') or 100,
-                                       grace_period=grace_period,
-                                       reduction_factor=3)
-
-        # Define the callbacks for the hyperparameter search
-        tuner_callbacks = [WandbLoggerCallback(project='YOLOv8-tune')] if wandb else []
-
-        # Create the Ray Tune hyperparameter search tuner
-        tuner = tune.Tuner(trainable_with_resources,
-                           param_space=space,
-                           tune_config=tune.TuneConfig(scheduler=asha_scheduler, num_samples=max_samples),
-                           run_config=RunConfig(callbacks=tuner_callbacks, local_dir='./runs'))
-
-        # Run the hyperparameter search
-        tuner.fit()
-
-        # Return the results of the hyperparameter search
-        return tuner.get_results()
+        self._check_is_pytorch_model()
+        from ultralytics.yolo.utils.tuner import run_ray_tune
+        return run_ray_tune(*args, **kwargs)
 
     @property
     def names(self):
